pip-services4-azure-python/requirements.txt
pytest
azure-functions >= 1.8.0, < 2.0.0

pip_services4_commons >= 0.0.1, < 1.0.0
pip_services4_components >= 0.0.1, < 1.0.0
pip_services4_config >= 0.0.1, < 1.0.0
pip_services4_container >= 0.0.1, < 1.0.0
pip_services4_data >= 0.0.1, < 1.0.0
pip_services4_observability >= 0.0.1, < 1.0.0
pip_services4_rpc >= 0.0.1, < 1.0.0